db/pyproject.toml
[project]
name = "db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "faker>=37.1.0",
    "influxdb>=5.3.2",
    "influxdb-client[ciso]>=1.48.0",
    "joblib>=1.5.0",
    "paho-mqtt>=2.1.0",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
]
